Move base classes to @zkdb/storage
