+++ packages/storage/src/common/const.ts
@@ -1,0 +1,37 @@
+// Schema metadata
+export const ZKDATABASE_SCHEMA_COLLECTION = '_zkdatabase_management';
+
+// Group
+export const ZKDATABASE_GROUP_COLLECTION = '_zkdatabase_group';
+
+// User -> Group mapping
+export const ZKDATABASE_USER_GROUP_COLLECTION = '_zkdatabase_user_group';
+
+// Permission
+export const ZKDATABASE_METADATA_COLLECTION = '_zkdatabase_permission';
+
+// System management
+export const ZKDATABASE_GLOBAL_DB = '_zkdatabase_metadata';
+
+// Database settings
+export const ZKDATABASE_DB_SETTING_COLLECTION = '_zkdatabase_db_setting';
+
+// Merkle tree collection
+export const ZKDATABASE_MERKLE_TREE_COLLECTION = '_zkdatabase_merkle_tree';
+
+// Proof queue
+export const ZKDATABASE_DB_QUEUE_COLLECTION = '_zkdatabase_proof_queue';
+
+// Proof storage
+export const ZKDATABASE_DB_PROOF_COLLECTION = '_zkdatabase_proof_storage';
+
+// Metadata collections
+export const ZKDATABASE_METADATA = [
+  ZKDATABASE_SCHEMA_COLLECTION,
+  ZKDATABASE_GROUP_COLLECTION,
+  ZKDATABASE_USER_GROUP_COLLECTION,
+  ZKDATABASE_METADATA_COLLECTION,
+  ZKDATABASE_MERKLE_TREE_COLLECTION,
+  ZKDATABASE_DB_QUEUE_COLLECTION,
+  ZKDATABASE_DB_PROOF_COLLECTION,
+];
