+++ packages/storage/src/model/abstract/general.ts
@@ -12,6 +12,8 @@ import {
   UpdateResult,
   DeleteOptions,
   DeleteResult,
+  FindOptions,
+  CountDocumentsOptions,
 } from 'mongodb';
 import ModelBasic from './basic';
 import logger from '../../helper/logger';
@@ -20,8 +22,8 @@ import logger from '../../helper/logger';
  * ModelGeneral was build to handle global metadata, this is mongodb general model and it have nothing
  * to do with zkDatabase or record of zkDatabase
  */
-export class ModelGeneral extends ModelBasic {
-  public static getInstance(databaseName: string, collectionName: string) {
+export class ModelGeneral<T extends Document> extends ModelBasic<T> {
+  public static getInstance<P>(databaseName: string, collectionName: string) {
     return new Proxy(new ModelGeneral(databaseName, collectionName), {
       get(target: any, prop: string) {
         if (typeof target.collection[prop] === 'function') {
@@ -34,49 +36,62 @@ export class ModelGeneral extends ModelBasic {
   }
 
   public async updateOne(
-    filter: Filter<Document>,
-    update: UpdateFilter<Document> | Partial<Document>,
+    filter: Filter<T>,
+    update: UpdateFilter<T>,
     options?: UpdateOptions
-  ): Promise<UpdateResult<Document>> {
+  ): Promise<UpdateResult<T>> {
     logger.debug(`ModelGeneral::updateOne()`, filter, update);
-    return this.collection.updateOne(filter, { $set: update }, options);
+    return this.collection.updateOne(filter, update, options);
+  }
+
+  public async updateMany(
+    filter: Filter<T>,
+    update: UpdateFilter<T>,
+    options?: UpdateOptions
+  ): Promise<UpdateResult<T>> {
+    logger.debug(`ModelGeneral::updateOne()`, filter, update);
+    return this.collection.updateMany(filter, update, options);
   }
 
   public async insertOne(
-    doc: OptionalUnlessRequiredId<Document>,
+    doc: OptionalUnlessRequiredId<T>,
     options?: InsertOneOptions
-  ): Promise<InsertOneResult<Document>> {
+  ): Promise<InsertOneResult<T>> {
     logger.debug(`ModelGeneral::insertOne()`, doc);
     return this.collection.insertOne(doc, options);
   }
 
   public async insertMany(
-    docs: OptionalUnlessRequiredId<Document>[],
+    docs: OptionalUnlessRequiredId<T>[],
     options?: BulkWriteOptions
-  ): Promise<InsertManyResult<Document>> {
+  ): Promise<InsertManyResult<T>> {
     logger.debug(`ModelGeneral::insertMany()`, docs);
     return this.collection.insertMany(docs, options);
   }
 
   public async findOne(
-    filter: Filter<Document>
-  ): Promise<WithId<Document> | null> {
+    filter?: Filter<T>,
+    options?: FindOptions
+  ): Promise<WithId<T> | null> {
     logger.debug(`ModelGeneral::findOne()`, filter);
-    return this.collection.findOne(filter);
+    return this.collection.findOne(filter || {}, options);
   }
 
-  public async find(filter?: Filter<Document>): Promise<Document[]> {
+  public async find(filter?: Filter<T>, options?: FindOptions) {
     logger.debug(`ModelGeneral::find()`, filter);
-    return this.collection.find(filter || {}).toArray();
+    return this.collection.find(filter || {}, options);
   }
 
-  public async count(filter?: Filter<Document>): Promise<number> {
+  public async count(
+    filter?: Filter<T>,
+    options?: CountDocumentsOptions
+  ): Promise<number> {
     logger.debug(`ModelGeneral::count()`, filter);
-    return this.collection.countDocuments(filter || {});
+    return this.collection.countDocuments(filter || {}, options);
   }
 
   public async deleteOne(
-    filter?: Filter<Document>,
+    filter?: Filter<T>,
     options?: DeleteOptions
   ): Promise<DeleteResult> {
     logger.debug(`ModelGeneral::deleteOne()`, filter);
@@ -84,7 +99,7 @@ export class ModelGeneral extends ModelBasic {
   }
 
   public async deleteMany(
-    filter?: Filter<Document>,
+    filter?: Filter<T>,
     options?: DeleteOptions
   ): Promise<DeleteResult> {
     logger.debug(`ModelGeneral::deleteMany()`, filter);
